logic for promotions now workingnicely, with only one loop to process all promotions

+++ src/main/java/io/nzbee/domain/bag/item/DroolsBagItemWrapper.java
@@ -16,7 +16,7 @@ public class DroolsBagItemWrapper {
 		return this.bagItem.getBagItem().getQuantity();
 	}
 	
-	public int getBagQuantity() {
+	public Long getBagQuantity() {
 		return this.bagItem.getBagItem().getBag().getTotalQuantity();
 	}
